Review models and modify repositories

+++ src/main/java/ca/sait/vezorla/model/PurchaseOrder.java
@@ -28,7 +28,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class PurchaseOrder {
 	@Id
-	private int poNum;
+	private Long poNum;
 	
 	@NotNull
 	private Date dateOrdered;
